requirements.txt
python-dotenv
fastapi
pydantic
uvicorn
python-multipart
qdrant-client
langchain-openai
langchain-community
langchain
unstructured